+++ hbs-replace.js
@@ -1,0 +1,8 @@
+
+Handlebars.registerHelper("replace", function (str, a, b) {
+    if(str && typeof str === "string") {
+      return str.split(a).join(b);
+    }
+  }
+);
+